atlas/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
einops>=0.6.0
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
pillow>=9.0.0
